requirements_txt.txt
# RealityCheck Coach Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Gemini API
google-generativeai==0.3.2

# Image Processing
Pillow==10.2.0

# Async Support
httpx==0.26.0
aiofiles==23.2.1

# Environment Variables
python-dotenv==1.0.0

# Data Validation & Serialization
python-multipart==0.0.6

# Optional: Redis for session storage (Phase 2+)
# redis==5.0.1
# aioredis==2.0.1

# Optional: Database (Phase 2+)
# sqlalchemy==2.0.25
# asyncpg==0.29.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing async endpoints
